% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianDFM.R
\docType{package}
\name{BayesianDFM}
\alias{BayesianDFM}
\title{A package for running a Bayesian DFM}
\description{
The BayesianDFM package provides relevant functions for running a Bayesian
Dynamic Factor Model.
}
\section{BayesianDFM functions}{

\itemize{
\item bvar_jeff
\item comp_f_state
\item draw_lam
\item draw_sig
\item get_ic
\item get_nowcast
\item multimove_gibbs
\item prepare_data
\item run_model
}
}

