% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\docType{data}
\name{inventory}
\alias{inventory}
\title{Descriptive Statistics of indicators.}
\format{
A data.frame with 23 rows and 5 variables:
\describe{
\item{key}{tbd}
\item{type}{tbd}
\item{freq}{tbd}
\item{mean}{tbd}
\item{sd}{tbd}
}
}
\source{
datastream
}
\usage{
inventory
}
\description{
A table containing relevant descriptive statistics of the UK indicators.
}
\keyword{datasets}
